Convert tests into integration tests
