+++ .github/workflows/rust.yml
@@ -27,7 +27,7 @@ jobs:
     - name: Check formatting
       run: cargo fmt --check
     - name: Check clippy
-      run: cargo clippy
+      run: cargo clippy --lib --tests
   regen:
     runs-on: ubuntu-latest
     steps:
