+++ src/lib.rs
@@ -79,23 +79,12 @@
     html_logo_url = "https://unicode-rs.github.io/unicode-rs_sm.png",
     html_favicon_url = "https://unicode-rs.github.io/unicode-rs_sm.png"
 )]
-#![cfg_attr(feature = "bench", feature(test))]
 #![no_std]
-
-#[cfg(test)]
-#[macro_use]
-extern crate std;
-
-#[cfg(feature = "bench")]
-extern crate test;
 
 use tables::charwidth as cw;
 pub use tables::UNICODE_VERSION;
 
 mod tables;
-
-#[cfg(test)]
-mod tests;
 
 /// Methods for determining displayed width of Unicode characters.
 pub trait UnicodeWidthChar {
